requirements.txt
torch>=2.0.0
diffusers>=0.19.0
transformers>=4.30.0
accelerate>=0.21.0
streamlit>=1.24.0
Pillow>=9.5.0
numpy>=1.24.0
pathlib>=1.0.1
tqdm>=4.65.0
huggingface-hub>=0.16.0